+++ src/components/App/CustomAppBar.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { AppBar, Button, Toolbar, Tooltip, Menu, useMediaQuery } from '@material-ui/core';
-import LoadSaveScheduleFunctionality from './LoadSaveFunctionality';
-import AccountBase from './AccountBase';
+import ScheduleLoginManager from './ScheduleLoginManager';
 import { Assignment } from '@material-ui/icons';
 import MenuIcon from '@material-ui/icons/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
@@ -49,8 +48,7 @@ const CustomAppBar = (props) => {
 
                 <div style={{ flexGrow: '1' }} />
 
-                <AccountBase />
-                <LoadSaveScheduleFunctionality />
+                <ScheduleLoginManager></ScheduleLoginManager>
 
                 <ConditionalWrapper
                     condition={isMobileScreen}
